lyra/requirements.txt
colorhash
alluka==0.1.1
attrs
black
hikari==2.0.0.dev109
hikari-tanjun==2.5.3a1
lyricsgenius==3.0.1
PyYAML
requests
ytmusicapi==0.20.0
python-dotenv
mutagen
pymongo
pymongo[srv]
colorama
Pillow
scipy
sklearn
numpy
soundcloud-lib
memoization
# ./lavasnek_rs-0.1.0_alpha.4-cp310-none-win_amd64.whl

uvloop
./lavasnek_rs-0.1.0_alpha.4-cp310-cp310-manylinux_2_24_x86_64.whl